Get phenophase information for a single plant

This function retrieves all phenophase data for a single plant_id. Can be subset to specific years, or specific start and end dates.

If shape = 'long' (the default) then columns will be: c('PLANT_ID','DATE','PHENOPHASE','STATUS','NOTE_FLAG','PHOTO_FLAG','SITE_CODE', 'SPP_CODE', 'FUNC_GRP_CODE')

If shape = 'wide' then all thephenophase codes will be columns.

get_plant_phenophase(
  plant_id,
  years = NULL,
  start_date = NULL,
  end_date = NULL,
  shape = "long"
)

Arguments

plant_id
string. unique plant identifier

years
Optional. integer or vector of integer for the years desired. years must be consecutive.

start_date
Optional. A string with format 'YYYY-MM-DD'. Get visit information from this date forward. Default is all prior dates.

end_date
Optional. A string with format 'YYYY-MM-DD'. Get visit information up to this date. Default is all dates up to todays date.

shape
string. 'wide' or 'long' for a data.frame in the respective format. default 'long'

Value

data.frame of phenophases by date

Examples

get_plant_phenophase('CRATCA01')

#> Error : Failed to connect: Can't connect to MySQL server on 'jornada-vdbmy.jrn.nmsu.edu' (110)

#> Error in db_connect(): Cannot connect to server: jornada-vdbmy.jrn.nmsu.edu, are you on the Jornada network?

get_plant_phenophase('CRATCA01', start_date = '2012-01-01', end_date = '2012-12-31')

#> Error : Failed to connect: Can't connect to MySQL server on 'jornada-vdbmy.jrn.nmsu.edu' (110)

#> Error in db_connect(): Cannot connect to server: jornada-vdbmy.jrn.nmsu.edu, are you on the Jornada network?