Get phenophase information for a single species.

This function retrieves all phenophase data for a single species. All individual at all sites for the species will be return. Can be subset to specific years, or specific start and end dates.

See the output from get_plant_info() for valid species codes.

If shape = 'long' (the default) then columns will be: c('PLANT_ID','DATE','PHENOPHASE','STATUS','NOTE_FLAG','PHOTO_FLAG','SITE_CODE', 'SPP_CODE', 'FUNC_GRP_CODE')

If shape = 'wide' then all thephenophase codes will be columns.

get_species_phenophase(
  spp_code,
  years = NULL,
  start_date = NULL,
  end_date = NULL,
  shape = "long"
)

Arguments

spp_code
string. unique species identifier

years
Optional. integer or vector of integer for the years desired. years must be consecutive.

start_date
Optional. A string with format 'YYYY-MM-DD'. Get visit information from this date forward. Default is all prior dates.

end_date
Optional. A string with format 'YYYY-MM-DD'. Get visit information up to this date. Default is all dates up to todays date.

shape
string. 'wide' or 'long' for a data.frame in the respective format. default 'long'

Value

data.frame of phenophases by date

Examples

get_species_phenophase('PRGL')

#> # A tibble: 277,080 x 11
#>    PLANT_ID DATE       NOTES_FLAG PHOTO_FLAG SITE_CODE SPP_CODE FUNC_GRP_CODE
#>    <chr>    <date>          <int>      <int> <chr>     <chr>    <chr>        
#>  1 CRPRGL01 2010-03-17          0          0 CR        PRGL     DS           
#>  2 CRPRGL01 2010-03-17          0          0 CR        PRGL     DS           
#>  3 CRPRGL01 2010-03-17          0          0 CR        PRGL     DS           
#>  4 CRPRGL01 2010-03-17          0          0 CR        PRGL     DS           
#>  5 CRPRGL01 2010-03-17          0          0 CR        PRGL     DS           
#>  6 CRPRGL01 2010-03-17          0          0 CR        PRGL     DS           
#>  7 CRPRGL01 2010-03-17          0          0 CR        PRGL     DS           
#>  8 CRPRGL01 2010-03-17          0          0 CR        PRGL     DS           
#>  9 CRPRGL01 2010-03-17          0          0 CR        PRGL     DS           
#> 10 CRPRGL01 2010-03-17          0          0 CR        PRGL     DS           
#> # ... with 277,070 more rows, and 4 more variables: DOY <dbl>, YEAR <dbl>,
#> #   PHENOPHASE <chr>, STATUS <int>

get_species_phenophase('PRGL', start_date = '2012-01-01', end_date = '2012-12-31')

#> # A tibble: 24,480 x 11
#>    PLANT_ID DATE       NOTES_FLAG PHOTO_FLAG SITE_CODE SPP_CODE FUNC_GRP_CODE
#>    <chr>    <date>          <int>      <int> <chr>     <chr>    <chr>        
#>  1 CRPRGL01 2012-01-03          0          1 CR        PRGL     DS           
#>  2 CRPRGL01 2012-01-03          0          1 CR        PRGL     DS           
#>  3 CRPRGL01 2012-01-03          0          1 CR        PRGL     DS           
#>  4 CRPRGL01 2012-01-03          0          1 CR        PRGL     DS           
#>  5 CRPRGL01 2012-01-03          0          1 CR        PRGL     DS           
#>  6 CRPRGL01 2012-01-03          0          1 CR        PRGL     DS           
#>  7 CRPRGL01 2012-01-03          0          1 CR        PRGL     DS           
#>  8 CRPRGL01 2012-01-03          0          1 CR        PRGL     DS           
#>  9 CRPRGL01 2012-01-03          0          1 CR        PRGL     DS           
#> 10 CRPRGL01 2012-01-03          0          1 CR        PRGL     DS           
#> # ... with 24,470 more rows, and 4 more variables: DOY <dbl>, YEAR <dbl>,
#> #   PHENOPHASE <chr>, STATUS <int>